Refactor: add view classes
