+++ src/main/java/io/output/EnterOutput.java
@@ -1,0 +1,13 @@
+package io.output;
+
+/**
+ * 맨 처음 화면<br/> 로그인, 회원가입 선택하는 문구 출력
+ */
+public class EnterOutput {
+
+    public void printInitView() {
+        System.out.println("[ 뇌정지 호텔 ]");
+        System.out.println("1. 로그인");
+        System.out.println("2. 회원가입");
+    }
+}
